Fixing the trap #9 issue for NCR710 ODE
Initially when we boot the HPUX ODE with the NCR710 Device model
We were facing the issue of:
SeaBIOS: Detected trap #9, at 0x0:0x4fadc, IIR=0x0, IOR addr=0x0:0x0
This was due to the fact that we have hardcoded the value of limit
transfer size based on scsi controller capability.

Now we have fixed it by properly adjusting for the smaller lasi size for
the NCR710 SCSI Controller.

Signed-off-by: Soumyajyotii Ssarkar <soumyajyotisarkar23@gmail.com>